+++ test/logger.js
@@ -1,35 +1,32 @@
-'use strict';
-
-const assert = require('assert');
-
-const { describe, it } = require('node:test');
-const logger = require('../lib/logger');
+import assert from 'node:assert';
+import { describe, it } from 'node:test';
+import * as logger from '../lib/logger.js';
 
 describe('logger', function () {
 
   ['info', 'error', 'debug'].forEach((level) => {
-    it('has ' + level + ' function', function () {
+    it('has ' + level + ' function', () => {
       assert.equal(typeof logger[level], 'function');
     });
   });
 
   describe('emits log entries', function () {
-    it('debug', function () {
-      process.env.DEBUG=1;
+    it('debug', () => {
+      process.env.DEBUG = 1;
       assert.ifError(logger.debug('test debug'));
       delete process.env.DEBUG;
     });
 
-    it('info', function () {
+    it('info', () => {
       delete process.env.NODE_ENV;
       assert.ifError(logger.info('test info'));
-      process.env.NODE_ENV='test';
+      process.env.NODE_ENV = 'test';
     });
 
-    it('error', function () {
+    it('error', () => {
       delete process.env.NODE_ENV;
       assert.ifError(logger.error('test error'));
-      process.env.NODE_ENV='test';
-    })
-  })
-})
+      process.env.NODE_ENV = 'test';
+    });
+  });
+});